refactor(fsd): partyroom, playlist 슬라이스 그루핑
