+++ src/features/edit-profile-meta/ui/v2.component.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { useState } from 'react';
-import V2EditMode from 'features/edit-profile-meta/ui/v2-edit.component';
-import V2ViewMode from 'features/edit-profile-meta/ui/v2-view.component';
+import V2EditMode from './v2-edit.component';
+import V2ViewMode from './v2-view.component';
 
 type ProfileEditFormV2Props = {
   onClickAvatarSetting?: () => void;
