+++ src/features/playlist/add/README.md
@@ -1,0 +1,1 @@
+playlist/add 와 playlist/search-youtube-music를 위젯에서 조합하여 사용
